feat(core): Keep track of loggers created through the default factory to allow out-of-order initialization of registry.
This can happen that an application starts retrieving loggers from the
registry before initializing it or configuring it (through a
configuration file).

Prior to this patch, we were not updating previously created loggers
after configuring the registry. We now keep track of loggers that have
been created through the default registry and update their sink and
level if needed.