core: Make sure to initialize the default logger
The default logger is never initialized, even if there is a default
logger factory. As a result, the basic logger example segfaults when
run. Note that by default, the default logger also doesn't have a sink,
so we have to set one up in the example to get any output.